pyproject.toml
[tool.poetry]
name = "PlayingPong"
version = "0.1.0"
description = "A Reinforcement Learning Approach to implementing automated game play of Pong using only user-viewable data."
authors = ["Sebastian Engels <me@sebastian-engels.com>"]
readme = "README.md"
packages = [{include = "PlayingPong"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
jupyter = "^1.0.0"
gymnasium = {extras = ["accept-rom-license", "atari"], version = "^0.29.1"}
numpy = "^1.26.3"
opencv-python = "^4.9.0.80"
tensorboard = "^2.15.1"
pytorch-ignite = "^0.4.13"
pettingzoo = {extras = ["accept-rom-license", "atari"], version = "^1.24.3"}
autorom = "0.4.2"
supersuit = "^3.9.2"
torchrl = "^0.2.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.11"
black = "^23.12.1"
isort = "^5.13.2"
mypy = "^1.8.0"
pre-commit = "^3.6.0"
pytest = "^7.4.4"

[tool.poetry.group.test.dependencies]
moviepy = "^1.0.3"
pygame = "^2.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
